aliased svim to preserve pythonpath for sudo
PYTHONPATH environment variable was not preserved for sudo, resulting in an
annoying message and not utilizing the powerline statusbar when editing a file
with vim. Using the -E option did not preserve all environment variables for
some reason. Thus, svim was aliased to specifically add the PYTHONPATH variable
for vim.